feat: Complete Task 01 and 02 - Move frontend GUI out of compilation (#60)
* feat: Complete Task 01 - Move frontend GUI out of compilation

- Archive all frontend components to src/_archived_frontend/
- Preserve backend libraries (OPDB, PinballMap, image-storage)
- Create minimal replacement frontend for development
- Update TypeScript and ESLint configs to exclude archive
- Complete Task 02 as side effect (Playwright tests archived)

Key changes:
- All React components moved to archive
- Frontend pages and routing archived
- Frontend-specific libraries archived
- Minimal layout/page/providers created
- Server-side query-client restored for SSR
- Build and TypeScript validation passing

Backend functionality preserved:
- All API routes functional
- Database operations intact
- External service integrations working
- Authentication system operational



* clean up newline

* Run CI on our branches

---------
