+++ .github/workflows/codeql.yml
@@ -5,7 +5,6 @@ on:
   push:
     branches: [ "main" ]
   pull_request:
-    branches: [ "main" ]
   schedule:
     - cron: '30 2 * * 1' # Run weekly
