+++ src/_archived_frontend/_components/auth-button.tsx
