+++ src/_archived_frontend/_components/issue-image-gallery.tsx
