+++ src/_archived_frontend/_components/issue-image-upload.tsx
