+++ src/_archived_frontend/_components/organization-logo-upload.tsx
