+++ src/_archived_frontend/_components/profile-picture-upload.tsx
