+++ src/_archived_frontend/_components/user-avatar.tsx
