+++ src/_archived_frontend/issues/[id]/page.tsx
