+++ src/_archived_frontend/location/[id]/page.tsx
