+++ src/app/page.tsx
@@ -1,51 +1,8 @@
-"use client";
-
-import { Container, Typography, CircularProgress, Alert } from "@mui/material";
-import React from "react";
-
-import { IssueSubmissionForm } from "~/app/_components/issue-submission-form";
-import { api } from "~/trpc/react";
-
-// Simplified homepage - only shows issue submission form
-
-export default function Home() {
-  // API queries - only need game instances for the issue form (public endpoint)
-  const {
-    data: gameInstances,
-    isLoading: isLoadingInstances,
-    error: instanceError,
-  } = api.gameInstance.getAllForIssues.useQuery();
-
-  // No mutations needed for the public issue form
-
-  const isLoading = isLoadingInstances;
-  const hasError = instanceError;
-
-  if (isLoading) {
-    return (
-      <Container maxWidth="lg" sx={{ py: 4, textAlign: "center" }}>
-        <CircularProgress />
-        <Typography variant="h6" sx={{ mt: 2 }}>
-          Loading...
-        </Typography>
-      </Container>
-    );
-  }
-
-  if (hasError) {
-    return (
-      <Container maxWidth="lg" sx={{ py: 4 }}>
-        <Alert severity="error">
-          Error loading data: {instanceError?.message}
-        </Alert>
-      </Container>
-    );
-  }
-
-  // Only show the issue submission form (besides the top bar)
+export default function HomePage() {
   return (
-    <Container maxWidth="sm" sx={{ py: 6 }}>
-      <IssueSubmissionForm gameInstances={gameInstances ?? []} />
-    </Container>
+    <main>
+      <h1>PinPoint Backend Refactor</h1>
+      <p>Frontend being rebuilt. Check back soon!</p>
+    </main>
   );
 }
