+++ tsconfig.json
@@ -38,5 +38,5 @@
     "**/*.js",
     ".next/types/**/*.ts"
   ],
-  "exclude": ["node_modules", "scripts/**/*.cjs"]
+  "exclude": ["node_modules", "scripts/**/*.cjs", "src/_archived_frontend/**/*"]
 }
